Change visibility of the verify methd for retry logic and moving helper method to another file
